refactor(async-upload): remap revived modules to post-reorg lib/db paths
