+++ frontend/tests/integration/ingest-batch.test.ts
@@ -62,7 +62,7 @@ const sharedState = vi.hoisted(() => ({
 // ConnectionManager mock — routes every DB call (including the stored procedure
 // CALL, withTransaction, and GET_LOCK application locking) to the shared real
 // MySQL connection so the production sequence runs against real MySQL state.
-vi.mock('@/config/connectionmanager', () => {
+vi.mock('@/lib/db/connectionmanager', () => {
   const manager = {
     executeQuery: async (query: string, params?: unknown[], transactionID?: string) => {
       if (!sharedState.connection) throw new Error('Test DB connection not initialized');
@@ -127,7 +127,7 @@ vi.mock('@/ailogger', () => ({
   }
 }));
 
-import ConnectionManager from '@/config/connectionmanager';
+import ConnectionManager from '@/lib/db/connectionmanager';
 import { stageMeasurementChunk, type StageMeasurementChunkParams } from '@/lib/uploads/stage-measurements';
 import { ingestBatch, IngestBatchAbortedError, type IngestBatchResult } from '@/lib/uploads/ingest-batch';
