+++ frontend/tests/integration/record-invalid-rows.test.ts
@@ -42,7 +42,7 @@ const sharedState = vi.hoisted(() => ({
 
 // ConnectionManager mock — routes every DB call to the shared real MySQL
 // connection so commits/rollbacks operate on actual transactions.
-vi.mock('@/config/connectionmanager', () => {
+vi.mock('@/lib/db/connectionmanager', () => {
   const manager = {
     executeQuery: async (query: string, params?: unknown[], transactionID?: string) => {
       if (!sharedState.connection) throw new Error('Test DB connection not initialized');
@@ -91,7 +91,7 @@ vi.mock('@/ailogger', () => ({
   }
 }));
 
-import ConnectionManager from '@/config/connectionmanager';
+import ConnectionManager from '@/lib/db/connectionmanager';
 import { recordInvalidRows, deleteUnresolvedRowsForBatch, type InvalidRowContext } from '@/lib/uploads/record-invalid-rows';
 
 // ---------------------------------------------------------------------------
